Add optional performance testing to test suite
If the PREFIXTREE_PERF environment variables is set, a performance test
will be run comparing PrefixTree to dict for CPU and memory consumption.
This is intended to catch regressions in performance.

Additionally, a new make target called 'performance' is added. This will
print the penalty of PrefixTree compared to dict.